src/requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-cors-headers==3.7.0
django-debug-toolbar==2.2
django-environ==0.8.1
django-extensions==3.1.3
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
generics==4.0.0
gunicorn==20.0.4
idna==3.2
image==1.5.33
importlib-metadata==1.6.0
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
more-itertools==8.2.0
oauthlib==3.1.1
packaging==21.3
Pillow==10.0.0
pluggy==1.0.0
psycopg2-binary==2.9.6
py==1.11.0
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==3.0.6
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.6.3
sqlparse==0.4.2
testfixtures==6.18.3
toml==0.10.2
tzdata==2021.5
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.1.9
webcolors==1.11.1
zipp==3.1.0
